+++ src/tree.rs
@@ -1,6 +1,6 @@
 use crate::{
     node::JsonTreeNode,
-    render_hooks::{RenderHooks, RenderKeyContext, RenderValueContext, ResponseContext},
+    render::{RenderHooks, RenderKeyContext, RenderValueContext, ResponseContext},
     value::ToJsonTreeValue,
     DefaultExpand, JsonTreeResponse, JsonTreeStyle,
 };
